add partage valeur thread joueurs

+++ common/src/message/challengedata.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug)]
 pub enum ChallengeData {
-    SecretSumModulo(u64),
+    SecretSumModulo(u128),
     SOS,
 }
