+++ common/src/message/hintdata.rs
@@ -4,6 +4,6 @@ use serde::{Deserialize, Serialize};
 pub enum HintData {
     RelativeCompass { angle: f32 },
     GridSize { columns: u32, rows: u32 },
-    Secret(u64),
+    Secret(u128),
     SOSHelper,
 }
